refactor: change folder structure and file names for api
